# apps/credential-manager/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/policies/check-reissue": {
      "post": {
        "operationId": "PoliciesController_checkAutoReissue",
        "summary": "Check if a credential can be reissued",
        "description": "Check if a credential can be reissued",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "query",
            "description": "Specifies the tenant ID",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PolicyParams" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The result of the policy evaluation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "result": { "type": "boolean" } }
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Tenant not found": {
                    "value": {
                      "statusCode": 404,
                      "message": "Tenant not found",
                      "data": null
                    }
                  },
                  "Credential not found": {
                    "value": {
                      "statusCode": 404,
                      "message": "Credential not found",
                      "data": null
                    }
                  }
                }
              }
            },
            "description": ""
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Internal server error": {
                    "value": {
                      "statusCode": 500,
                      "message": "Internal server error",
                      "data": null
                    }
                  }
                }
              }
            },
            "description": ""
          }
        },
        "tags": ["Policies"]
      }
    },
    "/v1/policies/check-revocation": {
      "post": {
        "operationId": "PoliciesController_checkAutoRevocation",
        "summary": "Check if a credential should be revoked",
        "description": "Check if a credential should be revoked",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "query",
            "description": "Specifies the tenant ID",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PolicyParams" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The result of the policy evaluation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "result": { "type": "boolean" } }
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Tenant not found": {
                    "value": {
                      "statusCode": 404,
                      "message": "Tenant not found",
                      "data": null
                    }
                  },
                  "Credential not found": {
                    "value": {
                      "statusCode": 404,
                      "message": "Credential not found",
                      "data": null
                    }
                  }
                }
              }
            },
            "description": ""
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Internal server error": {
                    "value": {
                      "statusCode": 500,
                      "message": "Internal server error",
                      "data": null
                    }
                  }
                }
              }
            },
            "description": ""
          }
        },
        "tags": ["Policies"]
      }
    },
    "/v1/policies/refresh": {
      "post": {
        "operationId": "PoliciesController_checkRefresh",
        "summary": "Check if a credential should be refreshed",
        "description": "Check if a credential should be refreshed",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "query",
            "description": "Specifies the tenant ID",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PolicyParams" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The result of the policy evaluation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "result": { "type": "boolean" } }
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Tenant not found": {
                    "value": {
                      "statusCode": 404,
                      "message": "Tenant not found",
                      "data": null
                    }
                  },
                  "Credential not found": {
                    "value": {
                      "statusCode": 404,
                      "message": "Credential not found",
                      "data": null
                    }
                  }
                }
              }
            },
            "description": ""
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Internal server error": {
                    "value": {
                      "statusCode": 500,
                      "message": "Internal server error",
                      "data": null
                    }
                  }
                }
              }
            },
            "description": ""
          }
        },
        "tags": ["Policies"]
      }
    }
  },
  "info": {
    "title": "Gaia-X OCM Credential Manager",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "PolicyParams": {
        "type": "object",
        "properties": {
          "credentialId": {
            "type": "string",
            "description": "The credential ID to check the policy for"
          }
        },
        "required": ["credentialId"]
      }
    }
  }
}